refactor: rename `objectHash` to `serialize`
