+++ test/ohash.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { objectHash, hash } from "../src";
+import { serialize, hash } from "../src";
 
 describe("hash", () => {
   it("hash", () => {
@@ -7,10 +7,10 @@ describe("hash", () => {
   });
 });
 
-describe("objectHash", () => {
+describe("serialize", () => {
   it("basic object", () => {
     expect(
-      objectHash({ foo: "bar", bar: new Date(0), bool: false }),
+      serialize({ foo: "bar", bar: new Date(0), bool: false }),
     ).toMatchInlineSnapshot(
       '"object:3:string:3:bar:string:24:1970-01-01T00:00:00.000Z,string:4:bool:bool:false,string:3:foo:string:3:bar,"',
     );
@@ -21,7 +21,7 @@ describe("objectHash", () => {
     form.set("foo", "bar");
     form.set("bar", "baz");
 
-    expect(objectHash(form)).toMatchInlineSnapshot(
+    expect(serialize(form)).toMatchInlineSnapshot(
       '"formdata:array:2:array:2:string:32:array:2:string:3:barstring:3:bazstring:32:array:2:string:3:foostring:3:bar"',
     );
   });
@@ -33,7 +33,7 @@ describe("objectHash", () => {
       }
     }
 
-    expect(objectHash(new Test())).toMatchInlineSnapshot(
+    expect(serialize(new Test())).toMatchInlineSnapshot(
       '"object:2:string:3:bar:string:3:baz,string:3:foo:string:3:bar,"',
     );
   });
